Add shared fixture to serve fonts
