+++ ui/playwright/index.js
@@ -1,8 +1,7 @@
 import { IconSets, OrIconSet, createMdiIconSet, createSvgIconSet } from "@openremote/or-icon";
 import { i18next } from "@openremote/or-translate";
 
-/// TODO: should not require the backend to be running for the iconset
-IconSets.addIconSet("mdi", createMdiIconSet("http://localhost:8080"));
+IconSets.addIconSet("mdi", createMdiIconSet(""));
 IconSets.addIconSet("or", createSvgIconSet(OrIconSet.size, OrIconSet.icons));
 
 await i18next
