+++ ui/test/index.d.ts
@@ -1,36 +1,6 @@
-import type { TestType, Locator } from "@playwright/experimental-ct-core";
 import type { Project } from "playwright/test";
+
 export * from "playwright/test";
-
-type ComponentProps<Component extends HTMLElement> = Partial<Component>;
-type ComponentSlot = number | string | ComponentSlot[];
-type ComponentSlots = Record<string, ComponentSlot> & { default?: ComponentSlot };
-
-type ComponentEvents = Record<string, Function>;
-
-export interface MountOptions<HooksConfig, Component extends HTMLElement> {
-  props?: ComponentProps<Component>;
-  slots?: ComponentSlots;
-  on?: ComponentEvents;
-  hooksConfig?: HooksConfig;
-}
-
-export interface MountResult<Component extends HTMLElement> extends Locator {
-  unmount(): Promise<void>;
-  update(options: {
-    props?: Partial<ComponentProps<Component>>;
-    slots?: Partial<ComponentSlots>;
-    on?: Partial<ComponentEvents>;
-  }): Promise<void>;
-}
-
-export const ct: TestType<{
-  mount<HooksConfig, Component extends HTMLElement = HTMLElement>(
-    component: new (...args: any[]) => Component,
-    options?: MountOptions<HooksConfig, Component>
-  ): Promise<MountResult<Component>>;
-}>;
-
 export { defineConfig, test } from "@playwright/test";
 export {
   defineConfig as defineCtConfig,
@@ -40,4 +10,6 @@ export {
 } from "@playwright/experimental-ct-core";
 
 export const createAppSetupAndTeardown: (app: string) => Project[];
+
+export * from "./fixtures/component";
 export * from "./fixtures/page";
